Update Stretch.java
Stretch:
Got rid of the boolean "dynamic". Whether volume means reps or seconds doesn't really mean anything coding-wise, it would only affect
toString() methods if anything.
The logic for the volume and intensity is nearly identical to Strength.
The concept of sets still makes sense for stretching, so I kept it.