Update Cardio.java
Made all constructors for each type of exercise use only 1 constructor
The logic required for each constructor (ie. rep values in Strength) should belong in a different method
The pseudo-code is explained in each class.

Reasoning for removing fixedVolume and fixedIntensity from Cardio:
If we want to allow the user to have a non-fixedVolume or non-fixedIntensity, will are required to make something like "int set"
for Cardio, which in my mind doesn't really make sense. For example, if the person wants to jog for 9 minutes and sprint for the last minute,
I think they are smart enough to do so instead of having two "sets" of running where they have one set that has a volume = 9 and intensity = moderate
then a "set" of volume = 1 and intensity = very_high. I really don't think it's necessary.